MCBayes

Documentation for MCBayes.

MCBayes.OnlineMoments

MCBayes.PSPoint

MCBayes.Stan

MCBayes.StepsizeConstant

MCBayes.StepsizeDualAverage

MCBayes.sample!

MCBayes.update!

MCBayes.Stan — Type

Stan(dims, chains, T = Float64; kwargs...)

Initialize Stan sampler object. The number of dimensions dims and number of chains chains are the only required arguments. The type T of the ...

Optionally, via keyword arguments, can set the metric, stepsize, seed, maxtreedepth, and maxdeltaH.

MCBayes.sample! — Function

sample!(sampler::Stan, ldg)

Sample with Stan sampler object. User must provide a function ldg(position; kwargs...) which accepts position::Vector and returns a tuple containing the evaluation of the joint log density function and a vector of the gradient, each evaluated at the argument position. The remaining keyword arguments attempt to replicate Stan defaults.

MCBayes.StepsizeDualAverage — Type

StepsizeDualAverage(initial_stepsize::Vector)

Construct a stepsize adapter using the dual averaging method by Nesterov 2009, as used in Stan. The length of initial_stepsize::Vector must be appropriate for the sampling algorithm for which this stepsize adapter will be used.

MCBayes.StepsizeConstant — Type

StepsizeConstant(initial_stepsize::Vector)

Construct a stepsize adapter for which the stepsize is fixed at it's initial value.

MCBayes.OnlineMoments — Type

OnlineMoments(d, c, update = true)

Returns an OnlineMoments struct with mean and variance Matrixes of size (d, c). When update!(om::OnlineMoments, x::Matrix) is called, update determines whether or not any updates will actually be applied.

MCBayes.update! — Function

Adam update.

update!(om::OnlineMoments, x::Matrix; kwargs...)

Update om's mean and variance Matrixes with the data contained in x. The rows of x and om.m (and thus om.v) must match. The columns of x and om.m must either match or om.m must have only 1 column. In the latter case, all columns of x will be used to update the same moments om.m and om.v.

MCBayes.PSPoint — Type

Point in general phase space.